+++ packages/terra-data-grid/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Changed
   * `boundingRef` is no longer necessary the section header's sticky behavior. The calculation now uses the table's container ref for calculating the offset.
 
+* Added
+  * Added `event` object to `onCellSelect` callback.
+
 ## 1.19.0 - (March 8, 2024)
 
 * Added