+++ packages/terra-table/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Changed
   * Table now uses its own container ref for calculating the offset while scrolling horizontally. `boundingRef` is no longer necessary for the scrolling behavior.
 
+* Added
+  * Added `event` object to `onCellSelect` callback.
+
 ## 5.12.0 - (March 8, 2024)
 
 * Changed